com_ecuser: coding convention

+++ 3.x/com_ecuser/admin/models/user.php
@@ -1,6 +1,9 @@
-<?php /** @package joomla.ecfirm.net
-* @copyright	Copyright (C) joomla.ecfirm.net. All rights reserved.
-* @license GNU General Public License version 2 or later. */
+<?php 
+/** 
+ * @package joomla.ecfirm.net
+ * @copyright	Copyright (C) joomla.ecfirm.net. All rights reserved.
+ * @license GNU General Public License version 2 or later.
+ */
 defined('_JEXEC') or die('Restricted access');
 
 
@@ -8,12 +11,17 @@ defined('_JEXEC') or die('Restricted access');
 class EcuserModelUser extends EcModelItemAdmin	{
 	
 	public function delete(&$valueKeys) { 
-		$valueKeys = (array)$valueKeys;
-		foreach($valueKeys as $valueKey) {
-			$ju = JFactory::getUser($valueKey);
-			if(!($ju->delete())) { $this->setError($ju->getError()); return false; } 
+		$bool = parent::delete($valueKeys);
+		
+		if($bool) {
+			$valueKeys = (array)$valueKeys;
+			foreach($valueKeys as $valueKey) {
+				$ju = JFactory::getUser($valueKey);
+				if(!($ju->delete())) { $this->setError($ju->getError()); return false; } 
+			}
 		}
-		return parent::delete($valueKeys);
+		
+		return $bool;
 	}
 	
 	/** Method to get article data.
@@ -23,6 +31,7 @@ class EcuserModelUser extends EcModelItemAdmin	{
 	public function getItem($keyValue = null)	{ 
 		$item = parent::getItem($keyValue); 
 		if(empty($item)) return $item;
+		
 		if($item->user > 0) {
 			$table = $this->getTable('User', 'JTable');
 			$table->load($item->user); //EcDebug::log($table);
@@ -30,6 +39,7 @@ class EcuserModelUser extends EcModelItemAdmin	{
 			$item->name = $table->name;
 			$item->email = $table->email; 
 		} //EcDebug::lp($item);
+		
 		return $item;
 	}
 	
@@ -40,11 +50,13 @@ class EcuserModelUser extends EcModelItemAdmin	{
 				|| ((is_string($value)) && (($value == ''))) || ($value == null))
 				unset($data[$key]); 
 		}
+		
 		$ju = JUser::getInstance($data['user']);//if user is zero then return new JUser
 		if(!($ju->bind($data))) { $this->setError('bind', $ju->getError()); return false; }
 		if(empty($ju->name)) $ju->name = $ju->username;
 		//if new then user id assigned
 		if(!($ju->save())) { $this->setError('save', $ju->getError()); return false; }
+		
 		if($data['user'] == 0) {
 			JUserHelper::addUserToGroup($ju->id, EcConst::USER_GROUP_REGISTERED);
 			return EcDml::insertRecord(array($this->name => $ju->id), $this->name);
